+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/RealmSessionStoreMigration.kt
@@ -43,7 +43,7 @@ import javax.inject.Inject
 class RealmSessionStoreMigration @Inject constructor() : RealmMigration {
 
     companion object {
-        const val SESSION_STORE_SCHEMA_VERSION = 10L
+        const val SESSION_STORE_SCHEMA_VERSION = 11L
     }
 
     override fun migrate(realm: DynamicRealm, oldVersion: Long, newVersion: Long) {
@@ -59,6 +59,7 @@ class RealmSessionStoreMigration @Inject constructor() : RealmMigration {
         if (oldVersion <= 7) migrateTo8(realm)
         if (oldVersion <= 8) migrateTo9(realm)
         if (oldVersion <= 9) migrateTo10(realm)
+        if (oldVersion <= 10) migrateTo11(realm)
     }
 
     private fun migrateTo1(realm: DynamicRealm) {
@@ -163,7 +164,7 @@ class RealmSessionStoreMigration @Inject constructor() : RealmMigration {
                 ?.addRealmListField(EditAggregatedSummaryEntityFields.EDITIONS.`$`, editionOfEventSchema)
     }
 
-    fun migrateTo9(realm: DynamicRealm) {
+    private fun migrateTo9(realm: DynamicRealm) {
         Timber.d("Step 8 -> 9")
 
         realm.schema.get("RoomSummaryEntity")
@@ -201,7 +202,7 @@ class RealmSessionStoreMigration @Inject constructor() : RealmMigration {
                 }
     }
 
-    fun migrateTo10(realm: DynamicRealm) {
+    private fun migrateTo10(realm: DynamicRealm) {
         Timber.d("Step 9 -> 10")
         realm.schema.create("SpaceChildSummaryEntity")
                 ?.addField(SpaceChildSummaryEntityFields.ORDER, String::class.java)
@@ -239,5 +240,11 @@ class RealmSessionStoreMigration @Inject constructor() : RealmMigration {
                 }
                 ?.addRealmListField(RoomSummaryEntityFields.PARENTS.`$`, realm.schema.get("SpaceParentSummaryEntity")!!)
                 ?.addRealmListField(RoomSummaryEntityFields.CHILDREN.`$`, realm.schema.get("SpaceChildSummaryEntity")!!)
+    }
+
+    private fun migrateTo11(realm: DynamicRealm) {
+        Timber.d("Step 10 -> 11")
+        realm.schema.get("EventEntity")
+                ?.addField(EventEntityFields.SEND_STATE_DETAILS, String::class.java)
     }
 }