+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/content/FileUploader.kt
@@ -31,12 +31,16 @@ import okhttp3.RequestBody.Companion.toRequestBody
 import okio.BufferedSink
 import okio.source
 import org.matrix.android.sdk.api.extensions.tryOrNull
+import org.matrix.android.sdk.api.failure.Failure
+import org.matrix.android.sdk.api.failure.MatrixError
 import org.matrix.android.sdk.api.session.content.ContentUrlResolver
+import org.matrix.android.sdk.api.session.homeserver.HomeServerCapabilities
 import org.matrix.android.sdk.internal.di.Authenticated
 import org.matrix.android.sdk.internal.network.GlobalErrorReceiver
 import org.matrix.android.sdk.internal.network.ProgressRequestBody
 import org.matrix.android.sdk.internal.network.awaitResponse
 import org.matrix.android.sdk.internal.network.toFailure
+import org.matrix.android.sdk.internal.session.homeserver.DefaultHomeServerCapabilitiesService
 import java.io.File
 import java.io.FileNotFoundException
 import java.io.IOException
@@ -46,6 +50,7 @@ import javax.inject.Inject
 internal class FileUploader @Inject constructor(@Authenticated
                                                 private val okHttpClient: OkHttpClient,
                                                 private val globalErrorReceiver: GlobalErrorReceiver,
+                                                private val homeServerCapabilitiesService: DefaultHomeServerCapabilitiesService,
                                                 private val context: Context,
                                                 contentUrlResolver: ContentUrlResolver,
                                                 moshi: Moshi) {
@@ -57,6 +62,22 @@ internal class FileUploader @Inject constructor(@Authenticated
                            filename: String?,
                            mimeType: String?,
                            progressListener: ProgressRequestBody.Listener? = null): ContentUploadResponse {
+        // Check size limit
+        // DO NOT COMMIT: 5 Mo
+        val maxUploadFileSize = 5 * 1024 * 1024L // homeServerCapabilitiesService.getHomeServerCapabilities().maxUploadFileSize
+
+        if (maxUploadFileSize != HomeServerCapabilities.MAX_UPLOAD_FILE_SIZE_UNKNOWN
+                && file.length() > maxUploadFileSize) {
+            // Known limitation and file too big for the server, save the pain to upload it
+            throw Failure.ServerError(
+                    error = MatrixError(
+                            code = MatrixError.M_TOO_LARGE,
+                            message = "Cannot upload files larger than ${maxUploadFileSize / 1048576L}mb"
+                    ),
+                    httpCode = 413
+            )
+        }
+
         val uploadBody = object : RequestBody() {
             override fun contentLength() = file.length()