+++ vector/src/main/java/im/vector/app/core/error/ErrorFormatter.kt
@@ -78,6 +78,9 @@ class DefaultErrorFormatter @Inject constructor(
                     throwable.error.code == MatrixError.M_LIMIT_EXCEEDED             -> {
                         limitExceededError(throwable.error)
                     }
+                    throwable.error.code == MatrixError.M_TOO_LARGE                  -> {
+                        stringProvider.getString(R.string.error_file_too_big_simple)
+                    }
                     throwable.error.code == MatrixError.M_THREEPID_NOT_FOUND         -> {
                         stringProvider.getString(R.string.login_reset_password_error_not_found)
                     }