+++ CHANGES.md
@@ -8,6 +8,7 @@ Improvements 🙌:
  - Add ability to install APK from directly from Element (#2381)
  - Delete and react to stickers (#3250)
  - Compress video before sending (#442)
+ - Improve file too big error detection (#3245)
 
 Bugfix 🐛:
  - Message states cosmetic changes (#3007)